course_app/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.11
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
function-logger==0.2
function-validator==0.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.0
MySQL-python==1.2.5
openapi-codec==1.3.2
passlib==1.7.1
pkg-resources==0.0.0
PyJWT==1.6.0
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
webargs==2.0.0

